created a common quick fix class
